Moved CacheSyncedCollection logic to JSONCollection
